pyproject.toml
[project]
name = "rag-benchmark-study"
version = "0.1.0"
description = "Modular RAG benchmark project with component-level evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.6.1",
    "faiss-cpu>=1.13.2",
    "kaleido>=0.2.1",
    "pandas>=2.0.0,<3",
    "plotly>=6.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.40.0",
    "tqdm>=4.67.3",
]

[project.scripts]
rag-baseline = "main:main"
